requirements.txt
PyQt6
Pillow
patool==2.3.0
filelock==3.15.4
jsonschema==4.22.0
# Test dependencies
pytest==8.2.0
pyfakefs==5.3.0
vulture==2.11
coverage==7.6.1
